backend/requirements.txt
matplotlib
uuid
google-cloud-core>=2.3.3
google-cloud-storage>=2.14.0
fastapi
pydantic
google-cloud-firestore
opencv-python-headless
numpy
pandas
seaborn
uvicorn
python-dotenv
scipy
google-generativeai
python-multipart
pytest
moviepy
requests
filterpy
torch
torchvision
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
flatbuffers==25.1.24
fonttools==4.55.8
google-api-core==2.24.1
google-auth==2.38.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.63.0rc1
h11==0.14.0
idna==3.10
jax==0.4.38
jaxlib==0.4.38
kiwisolver==1.4.8
# mediapipe==0.10.18  
ml_dtypes==0.5.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
scipy==1.15.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0